+++ src/Network/Network/Common/NetworkResourceManagerProfile.cs
@@ -1214,6 +1214,10 @@ namespace Microsoft.Azure.Commands.Network
                     .ForMember(
                         dest => dest.Active,
                         opt => opt.MapFrom(src => src.ActiveActive)
+                    )
+                    .ForMember(
+                        dest => dest.EnableHighBandwidthVpnGateway,
+                        opt => opt.MapFrom(src => src.EnableAdvancedConnectivity)
                     );
                 cfg.CreateMap<CNM.PSConnectionResetSharedKey, MNM.ConnectionResetSharedKey>();
                 cfg.CreateMap<CNM.PSConnectionSharedKey, MNM.ConnectionSharedKey>();
@@ -1232,6 +1236,7 @@ namespace Microsoft.Azure.Commands.Network
                         opt => opt.MapFrom(src => src.UseLocalAzureIpAddress)
                     );
                 cfg.CreateMap<CNM.PSIpsecPolicy, MNM.IpsecPolicy>();
+                cfg.CreateMap<CNM.PSTunnelConfig, MNM.VirtualNetworkGatewayConnectionTunnelProperties>();
                 cfg.CreateMap<CNM.PSVirtualNetworkGatewayIpConfiguration, MNM.VirtualNetworkGatewayIPConfiguration>();
                 cfg.CreateMap<CNM.PSTunnelConnectionHealth, MNM.TunnelConnectionHealth>();
                 cfg.CreateMap<CNM.PSVirtualNetworkGatewaySku, MNM.VirtualNetworkGatewaySku>();
@@ -1306,7 +1311,11 @@ namespace Microsoft.Azure.Commands.Network
                     .ForMember(
                         dest => dest.ActiveActive,
                         opt => opt.MapFrom(src => src.Active)
-                    );
+                    )
+                    .ForMember(
+                        dest => dest.EnableAdvancedConnectivity,
+                        opt => opt.MapFrom(src => src.EnableHighBandwidthVpnGateway)
+                     );
                 cfg.CreateMap<MNM.ConnectionResetSharedKey, CNM.PSConnectionResetSharedKey>();
                 cfg.CreateMap<MNM.ConnectionSharedKey, CNM.PSConnectionSharedKey>();
                 cfg.CreateMap<MNM.LocalNetworkGateway, CNM.PSLocalNetworkGateway>()
@@ -1324,6 +1333,7 @@ namespace Microsoft.Azure.Commands.Network
                         opt => opt.MapFrom(src => src.UseLocalAzureIPAddress)
                     );
                 cfg.CreateMap<MNM.IpsecPolicy, CNM.PSIpsecPolicy>();
+                cfg.CreateMap<MNM.VirtualNetworkGatewayConnectionTunnelProperties, CNM.PSTunnelConfig>();
                 cfg.CreateMap<MNM.VirtualNetworkGatewayIPConfiguration, CNM.PSVirtualNetworkGatewayIpConfiguration>();
                 cfg.CreateMap<MNM.TunnelConnectionHealth, CNM.PSTunnelConnectionHealth>();
                 cfg.CreateMap<MNM.VirtualNetworkGatewaySku, CNM.PSVirtualNetworkGatewaySku>();